fix(Farcaster): resolve circular dependency
