+++ src/requirements/Farcaster/components/FarcasterCastSmall.tsx
@@ -1,7 +1,7 @@
 import { Flex, HStack, Image, Link, Skeleton, Text } from "@chakra-ui/react"
 import { Chat, Heart, ShareNetwork, WarningCircle } from "@phosphor-icons/react"
 import DataBlock from "components/common/DataBlock"
-import { FarcasterCastData } from "./FarcasterCast"
+import { FarcasterCastData } from "../types"
 
 const FarcasterCastSmall = ({
   cast,
