+++ src/requirements/Farcaster/types.ts
@@ -1,0 +1,11 @@
+export type FarcasterCastData = {
+  hash: string
+  username: string
+  display_name: string
+  profile_pic: string
+  text: string
+  timestamp: string
+  likes: number
+  recasts: number
+  replies: number
+}
